Fuse.Drawing.Primitives: fix precision-issues in Circle renderer
While we're at it, add a test that verifies that a basic circle is
rendered correctly. This  doesn't cover the big circle-case that
the patch fixes, just makes sure we don't break small circles (which
is the most fragile bit here).

The reason we're not testing big circles, is that we would have to
make sure we had a FP16-fragment shader GPU we're running the tests
on, which we don't currently have. So it wouldn't actually pick up
any regressions.