Android: Fix TextView textAlignment rendering bug
When using a TextAlignment of Center or Right, rendering the TextView to a bitmap would end up rendering every character at the same location (on top of eachother). Turns out that if you use a TextAlignment of Center or Right the Width in pixels of the internal TextLayout would change to for example 1024 * 1024 instead of the width of the string. The origin of the string would also change to 1024 * 1024 - widthOfString. This changes the meaning of scroll position, a scroll position of 0 is not valid in a TextView with TextAlignment Center or Right. Setting target.setHorizontallyScrolling(source.getScrollX() > 0) makes the rendering TextView updates its internal layout state.

This lets us remove the previous alignment hack and now the scrollposition in the TextInput will be retained when switching between GL and Native.