Made ValueForwarder.IValueListener internal
The docs generator tried to create links to this interface because it is
implemented by some public classes (like `With`, although they are
explicitly implemented). I guess this should be considered a bug in the
docs generator, but if this is an OK change anyway, its an easier fix
for now.