Fuse.Elements: restore state when CompositEffects fail
A small detail from the stack-trace in fusetools/fuselibs#4211 is
that we seem to fail extra badly here, because of exception-unsafe
code.

Let's fix that part, as we can easily use try-finally for this.

There's no real changelog or docs needed here, as this just cleans
up the error-reporting when there's already something else gone
wrong; users shouldn't usually see this.