Fuse.Controls.Video: fix video-size for HLS-streams on iOS
We currently fail hard on HLS-streams on iOS, where we get a video
size of zero instead of the actual size.

As a half-measure, let's wait for the first resize-event until we
report the video as loaded. This allows us to correctly render HLS
streams that never change the resolution, which is better than
nothing.