Experimental.TextureLoader: downsample large bitmaps
Bitmaps that are larger than the maximum supported texture-size of
the GPU currently fails to load, producing an OpenGL error, and a
seemingly black texture.

To at least get something related to the image on-screen, let's
down-sample such images until they fit in a texture. We still lie
to the rest of the system about the texture-size later, so layout
etc will use the real size rather than the down-sampled size.

This is only implemented for C++-targets. That might be acceptable
for now, as all our mobile targets, which are the ones that are
likely to have low max-texture sizes. Modern desktop GPUs can do
16k x 16k textures, many mobile phones only support 4096 x 4096.