Fuse.Elements: Reduce batched rect size by a small margin
Since the batched rects are padded with a 1-pixel boundary that's
supposed to be empty, we can tighten up the draw rect by 0.5 pixels on
all sides and still be sure that we won't clip any of the content
inside. This is both a small optimization and also an additional safety
net if in the future we have more problems with batched elements drawing
outside of their allocated rects in the batch atlases.