add benchmark application
It's often useful to keep track of some performance cases over time,
so let's add a framework where it's easy to add cases and run them.

This is a pretty basic app, and we should probably need to both add
counters and tweak current functionality before this is very useful.
But at least it's a start.