Capture DrawRect vertices in world space, not clip space
This way we can properly composite rects that weren't drawn to the backbuffer. This also means we may not match up perfectly with any view/projection transforms the original draws did, but it seems like that won't really be a problem with most of the stuff we do in fuselibs.