make sure IncreaseFrameIndex happens right after draw
Doing things like InvalidateVisual in response to input requires that
the UpdateManager.FrameIndex was increased right after drawing, because
otherwise the frame index will be the same when the input happens,
causing elements that were validated the same frame to not be
invalidated.

This problem was introduced in 81eae3e, where another similar issue
was worked around. This should fix both cases.

Fixes issue #182