iOS: Make InputDispatch expand number of activePointers
There is no public API on iOS that will tell us how many simultanious touches are supported. When writing this code I believed that 10 fingers were the max, but when testing this on iPad I found that iPad supports 11 simultanious touches. Also according to the internet iPhone supports only 5 simultanious touches.

Refactoring to just use a list instead so that we can support an arbitrary number of active pointers