Fuse.Elements: do not use obsolete float.ZeroTolerance
float.ZeroTolerance has been marked as obsolete, so let's stop using
it directly. For now, let's just define it near the use-site as the
same value as it used to be.

Going forward, it could make sense to rethink these constants on a
case-by-case basis. But for now, this eliminates a bunch of warnings.