Android.TextEdit: fix TextEditRenderer alignment issues
Text alignment always broke the first time drawing due to the internal text alignment state in TextView not being valid. The previous workaround was to align the text ourselves with a grid. This aligned the text correct as long as it did not overflow its layout bounds. By reading the source code for TextView I figured out what to do to make the internal state update appropriately. See the code for more details.