Fuse.Scripting: merge all tests into one project
The CommonJS-tests are all doing the same thing, so there's no good
reason to compile the exact same C++ code over and over again just
to replace the JavaScript.

So let's bundle up everything in the same test, to reduce
compilation times during testing a bunch.

While we're at it, let's merge this with the last test-suite as well,
so we only have one project-file.

Before:

> $ time uno test -tmsvc Source/Fuse.Scripting/Tests/
>
> [snip]
>
> real    1m54.727s
> user    0m0.000s
> sys     0m0.030s

After:

> $ time uno test -tmsvc Source/Fuse.Scripting/Tests/
>
> [snip]
>
> real    0m16.375s
> user    0m0.030s
> sys     0m0.000s