Fuse.Scripting.V8: do not encourage use-after-free
It's nice to have methods that return a reference to the object it
works on so you can chain operations. However, release should never
be used like that, because it's effectively a free-method. Chaining
it like that will effectively be a use-after-free.

Luckily, nobody uses the return-value of the method, but let's just
get rid of it so nobody thinks this is a good idea in the future.