Fuse.Nodes: survive if there's no DataWatcher
This seems to happen during shutdown in some tests, if enough
objects are being disposed. So let's try to survive in this case.

This was noticed while adding some disposing to a test, so let's
add that here also, so we have a chance to notice similar regressions
in this functionality in the future.