iOS: Deal with pre-scaled CGContext in CanvasViewGroup
The surface api deals with scaling (PixelsPerPoint) but the CGContext that we get from UIView is already scaled with whatever is the current scale factor for it. This could lead to bad behavior if our PixelsPerPoint would differ from the UIView scale factor. We deal with this by removing the scale from the CGContext's transform