Fuse.Scripting.V8: do not release on the wrong context
Sometimes when releasing JS-values after a context has been disposed,
we get an AccessViolationException, probably due to using different
contexts (or rather, null-contexts) than the actual context used.

the `_context == null`-case seems to be a similarly mis-guided attempt
at surviving, where we just end up releasing in a collected context.

Since disposing sets the _context-member to the same value as the
`_context == null`-case fudged it with, these two cases should actually
be the same, and just as potentially crashy.

Let's instead assume that disposing a context correctly destroy all of
its resources. There's no way we can do any better, because the handle
to the context is lost at this point.

This fixes #237