Capture/render draw rects in GraphicsView.DrawFrame
This gives us visualization on platforms other than desktop. Since there's no stack, weird things will probably happen if there are multiple graphicsview's in play in the same app. For now I'm just going to ignore this and deal with it if we ever need to handle that case more thoroughly.