Implemented Visual.Children as an implicitly linked list.
This implementation avoids all memory allocation in connection with manipulating the Children list, and all the list operations are now O(1), including Contains(), InsertAfter() and Remove()

Updated all relevant call sites to use this API over the IList<T> interface (which is slower).

This is a big optimization over using MiniList, shaving 2-3 seconds of a case where 5000 text elements where added to a list.