Fuse.Nodes: unroot binding when removing
We unroot if there's only one, so it would be reasonable to expect
that we unroot if there's more.

This was found by code-inspection, and seems to have been like this
for a long time.

I also can't find a call-site for this, so it's probably never
triggered in practice.