Simplify rect capture conditionals
Now that we take in dc, we can do the target check in the Capture function. Additionally, this allows us to reformat the if statements so we get proper conditional compilation (something I hadn't realized until the compiler warned me it had corrected it for me in a few instances).