Fuse.Text: fix problem with ICU-clashing symbols
When using both -DUSE_HARFBUZZ and -DCOCOAPODS and some particular
pods, we end up linking to the wrong, AppStore banned symbols
instead of our bundled copy of ICU.

The approach of providing multiple copies of the same symbols, is
fragile, and some Cocoapods project re-writing ends up biting us
here. So let's get out of the "same symbols"-business, and use
ICUs built-in symbol-renaming mechanism instead.

Since we're using the system-provided copy of ICU on macOS, we
need to only do this for iPhone-builds.

The scripts to build and package ICU can be found here (sorry in
advance to people outside the fusetools organization, this repo
is currently private):

https://github.com/fusetools/uno-base/tree/build-icu