Wrap JS Date time value when marshalling to DateTime
Previously, this code would assume the result of getTime on a JS Date object would be a double. However, one critical value does not satisfy this assumption: the Unix epoch. This is because the result is 0, which is representable as an integer.

The fix is to wrap the value from getTime. We can assume the wrapped value is a double, so we cast to that from the resulting object, then cast its value to long.

I was sure I had tested this initially, and I probably did when writing the code, but since I wrote tests after, that case was likely not covered when adjusting the casting logic a bit while cleaning up. Whoops!