travis: install correct Mono version
Uno is pretty picky about the Mono-version we use, so let's make
sure we use the right one.

This is a bit sad, as it's way slower to install Mono manually, and
brew doesn't provide old Mono releases.

As a nice side-note, this was the last user of brew, which means
that we no longer need to update the brew database, which saves us
back some of the lost time.