Fuse.Elements: revert bad change
In 7f89944, we incorrectly assumed that this could never happen.
If you look at the callsite, you can see that this exact behavior
is expected.

This is not ideal behavior, but let's restore the old behavior for
now. We can rework this mechanism later for master.

While we're at it, add a test that picks up this breakage.

Fixes issue #375