Fix incorrect check for mono
Due to mono throwing a NotImplentedException when trying to set cache behavior,
the FuseJS/Http module disables caching for mono targets.

However, it only checks if we're running in preview,
resulting in a crash when running local dotnet build on osx,
as well as disabling cache on non-mono platforms that actually support it.