Fuse.Drawing.Surface: do nothing if bounds is empty
In Mono, System.Drawing is implemented using Cairo, and we seem to
get a crash when a clip-rectangle is empty in Cairo. So let's avoid
this.

The crash is currently not triggered due to a higher level check, but
that check is incorrect. So let's add these, and remove the higher
level check in the following commit.