Fuse.Elements: allow for some inaccuracy in test
AMD hardware is known to have a permanent bias to the texture
sampling location, probably to avoid some rounding issues when doing
nearest sampling. Unfortunately, this means our rendering will be a
bit blurry when going through a render-to-texture pass.

So let's add some error-tolerance here, it doesn't really affect
what the test tries to do; identify what kind of element is where.