ViewHandle.iOS: fix translation bug
The properties frame, bounds and center should all point to the same property behind the scenes. Looks like this is not the case in thirdparty libraries. For example setting a frame on Ad views from firebase has no effect on the view position.

Fixing this by always setting center to 0,0.