Fix #501
There was a problem with Video where if one used the `Play` action while
the video was still initializing, the action would not be applied.

After some conversing with Nekronos, we agreed that it would make more
sense for the video playback actions (Play, Pause, Stop) to (internally)
set a target state, so that if the user for example uses `Play` while
the Video is still initializing, the video will eventually start playing
when it is ready.