Fuse.Drawing.Primitives: avoid differing uniform precision
On Mali-400, the fragment-shader does not support highp floats, but
the vertex-shader does.

And since we set the default-precision to highp for all floats when
supported, we end up with different default-precision in the shader
stages. This cause issues when two uniforms have the same name in
both stages.

To side-step this, pre-calculate 1 / radius on the CPU, and use that
instead in the vertex-shader. This is a very minor optimization, but
is worth it simply due to avoiding using the same uniform-name.

Fixes fusetools/fuselibs#4266