Fuse.Drawing.Surface: re-introduce IsEmptyPath-hack
In 6bed36f, we dropped the incorrect IsEmptyPath-hack because we
had a more directed fix. However, this doesn't seem to cover all
cases. So let's reintroduce a less broken version of this check
(e.g one that checks if *both* width and height is zero) just to
be safe.

We should probably try to get rid of this at the benefit of a more
targetted fix later, but let's not try to deal with that right now.

This fixes a crash on startup in ManualTestingApp on dotnet-targets.