Fuse.Marshal: add a direct float -> float4 path
When reading this source, it's always a bit confusing that we don't
have a float -> float4 conversion-path. Instead, we convert to double
and then back to float4. This seems pointlessly complicated for such
a common case.

So let's add a direct code-path for float -> float4. This also allows
us to reuse some logic between the float and IArray.Length = 1 case
in the following patch.