Always compile ExifAndroidImplementation
Because of uno/issues/1368, the Android implementation was always compiled, even though it should have been stripped. Its gradle dependency was however not satisfied in this case, causing a java compiler error. The uxl file in this commit makes it so that the Exif android implementatino is always included in the build, as a workaround.