Fuse.Elements: use -1 for unspecified
In order to report an error when an exif-chunk exists, but does not
have a well-defined value, we need to tell the difference between
"no exif-orientation" and all orientation values. Luckily, the
orientation contains an unsigned value, so we can just reserve all
negative values for "unspecified".

This avoids a bunch of test-failures when we start using this data.