Fuse.Elements: drop return-value from Exif.FromByteArray
The whole design here was a bit mis-guided. Being able to get
exif-data isn't a all-or-nothing matter, so we need to be able
to deal with missing fields. The previous commit already does
this for Orientation, which is the only field we currently
support.

So let's simplify the interface a bit, and always return an
ExifData struct.

And since we always return some ExifData struct now, let's also
drop the helper, as it doesn't really do anything helpful.