Fuse.Reactive.JavaScript: fix no-toasts breakage
We used to report all diagnostics down to the tests, but a while
before 1.0, we stopped reporting temporal diagnostics. This means
that this was incorrect since then. So let's get rid of the code.

The code were never systematically tested, so this didn't do any
harm after all. But let's fix it now that we know.