hard-code FUSELIBS_NO_TOASTS on
FUSELIBS_NO_TOASTS is a blast from the past, when we needed the different
versions of the tool with different error-handling using the same fuselibs
version.

But that need is long gone, and the tool always enables this switch. So
let's get rid of the disabled case, as it just complicates things.