Fuse.Common: fix FontListParser for Android O
In Android O, FontVariationAxis was introduced to the fontlist format
in a way that wasn't backwards compatible with the old FontListParser.

Unfortunately, our copy has diverged quite a bit, and can't be replaced
with the new one, because the new one doesn't keep compatibility with
the old format.

So the best thing we can do for now, is to patch up the meat to match
the code better.

Sicne we don't really care about the FontVariationAxis-data yet, we
just ignore it to avoid depending on later SDK versions. We can revisit
this in the future.

This was not helped by #440, which just painted over the initial
problem instead of fixing it.

The change was adopted from the upstream, available here:
https://raw.githubusercontent.com/aosp-mirror/platform_frameworks_base/master/graphics/java/android/graphics/FontListParser.java