Make Invoke take an Action<Scripting.Context>
This is the start of a run of PRs to make the `Context` only be
accessible from the JS thread. The rest of the codebase will interface
through it via a (soon to be stripped down) `ThreadWorker`.

The way code will get access to the context is by running an action
on the JS thread. To enable this we start by making the `Invoke`
method take an `Action<Scripting.Context>` rather than an `Action`

There are two main evils in this commit:

- In Fuse.Reactive.JavaScript/Observable.uno:L93 we call Perform with
  null. This in temporary as Observable shouldnt be queryin if it is
  on the JS thread so this will be removed in a future commit.

- Adding Invoke to `IThreadWorker`. I didnt want to add more to this
  already questionable interface, however ultimately `Invoke` probably
  should be the only method on the interface. The others can be
  removed in a later PR