Move VMs into Fuse.Reactive.JavaScript
They are always all required by `Fuse.Reactive.JavaScript` so we are
merging them in and then (in the next commit) will rename them the
whole package to `Fuse.Scripting.JavaScript`. This then fits with the
pattern that `Fuse.Scripting` is generic across all scripting
languages and that `Fuse.Scripting.JavaScript` is the JavaScript
bindings for Fuse.