Example project: Using Path instead of Image to display Fuse logo
Because of a bug in the DotNet path renderer, which clipped a couple
pixels from the top part of the logo, i added `M -1.0 -10.0 l 1,1` to
the svg data, which pushes the bounds of the path a bit and avoids the
problem. Since we're not putting a stroke on the Path, the small extra
line is not visible.

This was only a problem in dotnet export.