*.Wrapping becomes TypeWrapper & *.ScriptClass merged into context
ThreadWorker.Wrapping is now a static helper class called TypeWrapper

All of ScriptClass' methods have moved to the context for now. This
showed lots of places that needed to take context explictly so we are
quickly cleaning up this references. I think I will move these out
into a static class ala TypeWrapper.

Sadly we also had to expose threadworker's context in an ugly way so
that it could be used by the tests. Those things are not threadsafe
and we really need to clean those up some day