Remove CanEvaluate
Instead we use whether the caller has obtained the Context to know if
they are on the correct thread or not.

The implications of this leaked all over the place which sucks, but
I'm thinking this is just our new use of Scripting.Context exposing
what we call from both thread. `ClearExclusive` for example looks like
it is used from both threads which sounds pretty data-racy to me.

Another concern is in `TreeArray`'s `ReplaceAllExclusive` method. I'm
not happy how ta.ReplaceAll is called from either thread however I
lack the insight to shore this up