IMirror.Reflect should run on JS Thread so require Scripting.Context
As with our other commits in this area this seeks to make the
Scripting functionality more robust through the use of
Scripting.Context.

This commit is interesting in that by updating the code we find an
exact place where the JS context is used incorrectly. Fixing this does
not belong in this commit but this should provide a sane place to
begin.