Fuse.Triggers: restore arg.sender argument on callbacks
Users depend on args.sender in the JavaScript handler of a Callback
to contain the name of the visual that was for instance pressed.

However, 51ec407 changed the sender to be the trigger itself, which
broke this behavior.

Let's restore this behavior, by lying about what the sender is inside
Callback. This should be the only user-visible effect of the
sender-change.

Fixes #659