Fuse.Elements: ignore zero-sized display-hints
This cause problems during testing, where we don't always have a
display set up.

While we're at it, throw an exception if we somehow manage to get
zero-sized Texture2D.MaxSize, which would mean this was queried
without a current OpenGL context. This has not been observed to
trigger, so it's purely a safety-mechanism.