Fuse.Elements: unpin CacheFramebuffers on Exceptions
This code wasn't very exception-safe, and leaks pins on
CacheFramebuffers on errors. Let's clean up this by handling the errors
and explicitly unpinning.

In the future, it might be an idea to use an IDisposable + using pattern
to avoid having to do this manually. This is however a bit easier said
than done, as we need to pin a variable amount of CacheFramebuffers in
some cases. So let's just put a band-aid on here for now.

This was noticed while locally introducing a bug that made the device
run out of memory during rendering, causing test-failures claiming we
were unpinning while going to the background.