Fuse.Elements: do not create crazy big element-atlases
During testing, we hit the case where we have no SystemUI-frame,
leading to the creation of max-texture-sized framebuffers. That's
not good on GPUs that support 16k x 16k textures, leading to quickly
running out of memory.

For desktop GPUs, we already limit things at 2k x 2k, which seems
like a reasonable limit here. So let's generalize that limit.