Move FuseJS field from ThreadWorker to JSContext
The FuseJS.Builtins is constructed with a specific context and so is
tied to it. This means it is only safe to use from that context. Due
to this we move it to the context and make it non-static.

Also we are trying to make the thread worker more focused on the
thread communication so fits with that goal.