Fuse.Scripting.JavaScript: throw all exceptions to the UI-thread
Throwing only the last exception from the JS-thread is inheretly racy,
as these threads are asynchronous. It's also not very helpful, as it
can hide what lead to the problem, and because of the racyness, this
might not be easy ti reproduce.

So, instead of doing a once-per-frame check for exceptions, let's queue
up an action-closure per exception thrown, so we throw all exceptions, in
the order they were thrown.

While we're at it, switch to using UpdateManager.PostAction, which is
actually thread-safe.

Fixes #705.