Fuse.Scripting.JavaScript: clean up JSContext lifetime
We no longer need to have this as a member, so let's just make it
an argument to RunInner, so we can use a using-statement instead.

This has a tiny benefit of also dispatching exceptions during
JSContext.Dispose(), which would otherwise have gone unhandled. Not
that anyone is going to care.