Fuse.Scripting.JavaScript: hook up OnTerminating in ctor
Yes, this is kind of the inverse of e10d463, but the whole
bootstrapping has changed since then, and this can't be null any
more. In fact, it's a static handler now.

But even more importantly, hooking this up from the JS thread isn't
thread-safe, so let's make sure we do it from the UI-thread.

The old code would also crash on shutdown if we didn't get far enough
during init to hook up the event.