Don't read TreeObservable state on JS thread
TreeObservable.Operation now traverses the graph from the UI thread when
the operation is performed, instead of on the JS thread when it is scheduled.

This ensures that the previous operation has completed before the next
operation is performed, so that the path of the next operation is valid.

Fixes #732