Fuse.Scripting.JavaScript: do not make mirror contain JS objects
When we truncate while reflecting the JS state, we accidentally left
JS objects in the mirror, because we weren't able to tell the difference
between truncation and primitive-types. So let's rewrite the code a bit
to be able to tell the difference.

Because the mirroring gives up after 50 recursions, make the
RouterRequest validation give up after 49 instead, so we retain the
same errors even when truncating.