clean up packages/projects/usings
There's two things going on here:
- reduction of using-statements
- switching from package-references to project references.

The former should in theory improve complilation-speeds a tad.

The latter is about correctness. If the user has a newer fuselibs
package installed than the in-tree version, the uno package manager
will try to resolve this with the newer one, which is not what we
want. So internal references in the fuselibs-repo should always be
project-references, not package-references.

This is a big, nasty commit, but untangling this without breaking
rebaseablitly is silly much work for little gain, so I'd prefer to
do it in one go.