Fuse.Scripting.JavaScript: dispose Worker when the last JavaScript-tag gets unrooted
This will force builtins to be reevaluated, which fixes some problems
with Polyfills/Window not being available when an app that uses
the JavaScript-tag, but *not* FuseJS gets rooted before one that uses
both.

This fixes problems with tests using AllTests.unoproj on feature-Models,
but the problem should also be possible to trigger from preview or other
carefully crafted apps.