Fuse.Reactive: drop context-versions of ISubscription
All call-sites can do better directly, due to internal knowledge
about the objects.

For instance, Fuse.Scripting.JavaScript.ObservableProperty knows
exactly the type of Subscription-object it gets, and can call
specific script-shortcut versions of the methods.

And Fuse.Scripting.JavaScript.TreeArray is going to dispatch to the
UI-thread anyway, so it can just do this a bit earlier.

This should fix some problems in premium packages that use this type.