Ignore all user properties with leading underscore
Revert "Ignore all user properties with leading underscore"

This reverts commit c5a1469768957ca9e41fa133d913cde17f0fc841.

We actually track wether something has changed to determine
if we should re-evaluate computed properties.
Since we're diffing with the node, and we're not setting private
properties on the node, it doesn't detect that a private field has
changed, and thus, no computed props are re-evaluated.

This is totally fixable, but not without significant refactoring.

Revert "Revert "Ignore all user properties with leading underscore""

This reverts commit dc0f54c7e9d500e1b5042a74cd19651c2710cd13.

Actually, we'll want to introduce this feature later on.
To not cause a regression in this case, we'll instead document it
as having no guarantees wrt. change detection.