Dispatch Scripting Tests to JS Thread
In this commit we have done some work to make JS tests safer. We were
seeing a lot of places where the tests were creating a tonne of
contexts and using them from various threads. We would rather start
& use the context in a way that matches our runtime behaviour.

These changes fix the crashes we were getting in tests when building
for JSCore on android (which, at the version we are using, is not
thread safe)

Changes are:
- Make JSTest: This dispatches tests to the JS thread, handles waiting
  on results and also injects the 'test' method previously provided by
  JSTestRecorder
- Remmove JSTestRecorder
- Update all problematic tests to use JSTest