'Fix' hang in ScriptPromise tests
Previous there were 3 promise tests sharing the same test panel.

the panel has a javascript tag

the javascript tag would, at module eval time, try and dispatch some
promises.

.e.g this top level expression `customPanel.getString().then(...)`

however the test only populated the promise it was interested in, so
I believe the other expressions would try and call `then` on null.

This seems to leave the js in a rather undefined that so that a few
times in a hundred the promise never resolves and we get a hang
resulting in the tests timing out.

By refactoring these tests to each own their own panel and
bootstrapping we avoid the issue.

However the lack of an error is a deeper issue that is being tracked
here: https://github.com/fusetools/fuselibs-public/issues/840