Fuse.Video.CIL: Build separately for Mac and Win
Create separate configurations for building the Fuse.Video.CIL C#
project on Mac and Windows, and load the right assemblies in Uno
depending on the platform.

This allows us to conditionally include only the projects that are
needed for the specific platform depending on the configuration we're
building, and also means that the Platform detection module is not
needed anymore since it's all statically determined.

This fixes an issue where the Fuse.Video.CIL assemblies failed to load
on Mac with recent versions of Uno because it tried to load
Windows-specific assemblies that weren't there.