Model: Detach from child at specific key
Every meta node keeps a list of its parents.
Each entry in this list contains:

    - The key that the child is attached to the parent by
    - The meta node of the parent

Previously, when detaching a child, we would remove whatever
relation to the parent appeared first in this list - completely ignoring the key.
This changes the removeAsParentFrom function to
accept the key to detach from in addition to the parent node.

Additionally, this introduces some stricter/more correct equality checks
when searching for the entry to remove from the list of parents.