Make stuff/uno pack use same versioning scheme
This should change the versioning of artifacts built with `stuff pack`
to be the same as for artifacts built with `uno pack`. In particular,
`AutomaticTestApp.stuff` and `ProjectTemplates.stuff` will now get the
same version numbers as `fuselibs.packages`. This will make things less
confusing for consumers of these packages.