Fuse.Drawing.Surface: hotfix for crash in StrokePathLinearGradient
On .NET (windows), we currently get a crash on start-up in
ManualTestingApp, due to an System.OutOfMemoryException inside the
LinearGradientBrush constructor. This happens when the start and end
points of the gradient are the same.

This is already a degenerate case, do whatever we end up doing in this
case is not important. We just don't want to crash.

We already have some code to deal with this, and this can be extended
to produce a single-pixel wide, diagonal gradient instead.