Model: Fix parent refs on array manipulation
Whenever an element is inserted or removed from an array,
the indices of later elements are also adjusted.
However, for every node in the model graph, we also keep track of which
nodes have a property that refers to this node. These parent references
were not adjusted accordingly.

Now, when we do an `insertAt` or `removeRange`, the parent references
of elements that appear after the index we insert/remove at will be
adjusted accordingly.