Model: Fix pushing an array onto an array
`Array.prototype.concat()` is recursive up to one level.
This means that if one of the arguments is an array, then its elements,
rather than the array itself, will be appended to the original array.
This changes it to always wrap the value in a single-element array,
which gives us the intended behavior.