Hide unhandled exception methods
These were never meant to be exposed to users, but it seems users are finding this and asking how to use it. So let's hide it so this doesn't happen again in the future. Note that we still want to keep the documentation for our own internal information, as well as the tool's.