Fuse.*Notifications: clarify ambiguous symbol
Both of these files uses both Uno.UX and Fuse.Reactive (which both
provides a Resource class), but neither clarify which one they want to
use. This leads to a compilation error on usage.

So let's fix it, by adding the fully qualified name here.

How this has gone unnoticed for this long is beyond me. I guess nobody
is using this functionality?

This was extracted out of #1197, but another case was added as well.